Localize calculation off cell width/height

+++ src/Painter.hpp
@@ -13,14 +13,14 @@ public:
     Painter(double scale_x, double scale_y);
 
     void CalcCellInfo();
-    double GetCellWidth() const { return _cell_width; }
-    double GetCellHeight() const { return _cell_height; }
+    int GetCellWidth() const { return _cell_width; }
+    int GetCellHeight() const { return _cell_height; }
     void Paint(SDL_Surface *, std::string_view text, const HlAttr &, const HlAttr &def);
 
 private:
     double _scale_x = 1.0;
     double _scale_y = 1.0;
-    double _cell_width = 0.0;
-    double _cell_height = 0.0;
+    int _cell_width = 0.0;
+    int _cell_height = 0.0;
     PtrT<PangoFontDescription> _font_desc;
 };
